Add functions & tests for getting monsters' stat info
Since this is the first R code added, these are the first tests also, so
the setup for that was done in this commit. The functions get info on
baseline stats or stat gain information for monsters using the monsters'
types, so that users don't have to manually filter the included dataset
(and because the form it's returned in will be useful internally).